Rename html to toHtml
